- XG state parameter observer pattern correction.
git-svn-id: http://svn.code.sf.net/p/qxgedit/code_attic@16 eb8aa3ee-cedb-4360-ae73-c3cf0b1998c6

+++ src/XGParamWidget.cpp
@@ -56,7 +56,7 @@ XGParamWidgetMap::~XGParamWidgetMap(void)
 
 // Add widget to map.
 void XGParamWidgetMap::add_widget (
-	QWidget *widget, XGParamMap *map, unsigned id )
+	QWidget *widget, XGParamMap *map, unsigned short id )
 {
 	XGParamInst inst(map, id);
 	m_widget_map.insert(widget, inst);
